+++ lib/audio_codec.dart
@@ -6,5 +6,5 @@ library;
 // export 'src/mpeg/mpeg_decoder.dart' show MpegDecoder;
 
 export 'src/flac/flac_decoder.dart' show FlacDecoder;
-export 'src/flac/flac_encoder.dart' show FlacEncoder, FlacEncoderConfig;
+export 'src/flac/encoder/flac_encoder.dart' show FlacEncoder, FlacEncoderConfig;
 export 'src/ogg/ogg_decoder.dart' show OggDecoder, OggAudioCodec;
